requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-pinecone>=0.0.3
flask
pypdf
python-dotenv
-e .